Move OffsetArray definition to TestHelpers
(cherry picked from commit e95b5e20c4e7d6abfd4345d9dc5a114c6153ffa6)
ref #17919
